pyproject.toml
[tool.poetry]
name = "zhiweitong"
version = "0.1.0"
description = "智维通数字员工平台 — 按 OpenCLAW 规范实现的 asyncio 核心（全新工程，拼音标识 zhiweitong）"
readme = "README.md"
authors = []
packages = [
    { include = "core" },
    { include = "shared" },
    { include = "config" },
    { include = "cli" },
    { include = "skills" },
]

[tool.poetry.dependencies]
python = "^3.11"
pydantic = "^2.0"
sqlalchemy = { extras = ["asyncio"], version = "^2.0" }
aiosqlite = "^0.20.0"
redis = "^5.0"
httpx = "^0.27.0"
# Typer <0.18 breaks with Click 8.3+ (options ignored → "unexpected extra argument" in CI).
typer = { extras = ["all"], version = "^0.18.0" }
anytree = "^2.12.0"
chromadb = "^0.5.0"
python-dotenv = "^1.0.0"

[tool.poetry.group.dev.dependencies]
pytest = "^8.0.0"
pytest-asyncio = "^0.24.0"
coverage = "^7.0.0"

[tool.poetry.scripts]
zhiweitong = "cli.main:run"

[tool.pytest.ini_options]
# 异步测试使用 tests 内 asyncio.run；若改用 pytest-asyncio 可配置 asyncio_mode

[tool.coverage.run]
source = ["core", "shared", "config", "cli", "skills"]
omit = ["*/tests/*"]

[tool.coverage.report]
fail_under = 0
# 门禁由 CI / make verify 显式执行：core/* ≥85%，Phase 1.4 skills/quick_consumption ≥90%；见 .github/workflows/ci.yml

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
